+++ src/components/features/settings/addon-verification/steps/ClaimAddonsStep.tsx
@@ -3,7 +3,7 @@ import { Button } from '@/components/ui/button';
 import { RefreshCcw, CheckCircle2, AlertCircle } from 'lucide-react';
 import type { PushChangesProps } from '../types';
 import { useUserStore } from '@/api/stores/userStore';
-import { useUpdateAddon } from '@/api/endpoints/useAddons';
+import { useUpdateAddon } from '@/api/appwrite/useAddons';
 import { LoadingSpinner } from '@/components/loading-overlays/LoadingSpinner';
 import TimerProgress from '../components/TimerProgress';
 import AddonFetcher from '../components/AddonFetcher';
